refactor: Car를 인스턴스로 갖는 CarList 클래스를 구현하여 리팩토링
